refactor, move deserializer logic to own files
