+++ Assets/Scripts/RemoteData/SceneData.cs
@@ -15,16 +15,6 @@ namespace RemoteData
         public string type;
         public Metadata metadata;
         public string version;
-
-        public static SceneData FromJson(string jsonString)
-        {
-            var settings = new JsonSerializerSettings
-            {
-                Converters = new List<JsonConverter> { new Vector2IntConverter() }
-            };
-            
-            return JsonConvert.DeserializeObject<SceneData>(jsonString, settings);
-        } 
         
         public class ContentData
         {
@@ -85,25 +75,5 @@ namespace RemoteData
                 public Vector3 cameraTarget; 
             }
         }
-        
-        
-        
-        
-        private class Vector2IntConverter : JsonConverter<Vector2Int>
-        {
-            public override Vector2Int ReadJson(JsonReader reader, Type objectType, Vector2Int existingValue, bool hasExistingValue, JsonSerializer serializer)
-            {
-                string[] parts = ((string)reader.Value).Split(',');
-                int x = int.Parse(parts[0]);
-                int y = int.Parse(parts[1]);
-                return new Vector2Int(x, y);
-            }
-
-            public override void WriteJson(JsonWriter writer, Vector2Int value, JsonSerializer serializer)
-            {
-                string stringValue = value.x + "," + value.y;
-                writer.WriteValue(stringValue);
-            }
-        }   
     }
 }
